i am fucking FURIOUS FUCK MINIFYCSS
for SOME reason it stripped out the `%` in `initial-value`, so the property declaration was invalid and nothing fucking worked. this was only an issue in prod and worked fine in dev. i hate this shit i hate this shit i hate this shit i hate this shit i hate this shit i hate this shit